chore: abstracted markWalletsAsInstalled util
